move aes encryption under network session
